fix: issue

+++ src/network/session/network_session.cpp
@@ -62,11 +62,11 @@ cyrex::util::UUID randomUUID()
 {
     std::random_device rd;
     std::mt19937_64 gen(rd());
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
+    std::uniform_int_distribution<unsigned int> dist(0, 255);
 
     std::array<uint8_t, 16> bytes{};
     for (size_t i = 0; i < 16; ++i)
-        bytes[i] = dist(gen);
+        bytes[i] = static_cast<uint8_t>(dist(gen));
 
     bytes[6] = (bytes[6] & 0x0F) | 0x40;
     bytes[8] = (bytes[8] & 0x3F) | 0x80;
